organize icons + add more icons
