+++ app/components/common/icons/AvatarImg.js
@@ -1,0 +1,11 @@
+export default function AvatarImg(props) {
+  return (
+    <div>
+      <img
+        class="w-12 h-12 mb-3 mr-3 rounded-full sm:mb-0"
+        src={props.imgURL}
+        alt={props.title}
+      />
+    </div>
+  );
+}
