+++ app/components/common/icons/stars/StarWhite.js
