+++ app/components/common/icons/stars/StarYellow.js
